Implemented The Chosen One Axe
Edited class files in RoleGame.java and WeaponFactory.java to
reflect implementation of new weapon.

Added files TheChosenOneAxe.java and ChosenAxeFate.java.

Class TheChosenOneAxe creates an instance of a Chosen One Axe.
It uses the class ChosenAxeFate to find the line of the file from
which it draws its damage value (thechosenone.txt).